requirements.txt
ipython==8.25.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
matplotlib==3.9.0
matplotlib-inline==0.1.7
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.34.0
pandas==2.2.2
PyPDF2==3.0.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.1
scipy==1.14.0
tiktoken==0.7.0
